feat(notification): 알림 설정을 우측 패널로 이동 및 라우트 제거
